Cashflows.Cashflow

Description

This module defines a cashflow

Documentation

class Cashflow a where

Cashflow type class

Methods

amount :: a -> Double

amount that is received (usually positive) or paid (usually negative)

payDate :: a -> Date

date on which the amount is paid

earlierThan :: a -> a -> Bool

True if first cashflow occurs on a date earlier than the second cashflow

data SimpleCashflow

Simple cash flow given by amount and pay date

Constructors

SimpleCashflow

Fields

scAmount :: Double

amount

scPayDate :: Date

pay date