Hql.Time.DayCounter

Description

This module defines day counters and corresponding day count and time measuring functions.

Documentation

data DayCounter

Day counter definitions

Constructors

Actual360
Actual/360 day count convention, also known as "Act/360", "A/360"

Thirty360
30/360 day count convention The 30/360 day count can be calculated according to US, European, or Italian conventions.

US (NASD) convention: if the starting date is the 31st of a month, it becomes equal to the 30th of the same month. If the ending date is the 31st of a month and the starting date is earlier than the 30th of a month, the ending date becomes equal to the 1st of the next month, otherwise the ending date becomes equal to the 30th of the same month. Also known as "30/360", "360/360", or "Bond Basis"

European convention: starting dates or ending dates that occur on the 31st of a month become equal to the 30th of the same month. Also known as "30E/360", or "Eurobond Basis"

Italian convention: starting dates or ending dates that occur on February and are grater than 27 become equal to 30 for computational sake.

Thirty360USA

Thirty360BondBasis

Thirty360European

Thirty360EurobondBasis

Thirty360Italian

Actual365Fixed
Actual/365 (Fixed) day count convention "Actual/365 (Fixed)" day count convention, also know as "Act/365 (Fixed)", "A/365 (Fixed)", or "A/365F".

According to ISDA, "Actual/365" (without "Fixed") is an alias for "Actual/Actual (ISDA)" (see ActualActual.) If Actual/365 is not explicitly specified as fixed in an instrument specification, you might want to double-check its meaning.

dayCount :: DayCounter -> Date -> Date -> Int

Number of days w.r.t. a given day counter for a given period.

yearFraction :: DayCounter -> Date -> Date -> Maybe Date -> Maybe Date -> Double

Yearfraction w.r.t. a given day counter and an optional reference period. The latter is only relevant for some day counters such as Act/Act.