Time.Calendar

Description

This module contains holiday calendar definitions and functions to apply adjustments to dates

Documentation

data BusinessDayConvention

These conventions specify the algorithm used to adjust a date in case it is not a valid business day.

Constructors

Following
Choose the first business day after the given holiday.

ModifiedFollowing
Choose the first business day after the given holiday unless it belongs to a different month, in which case choose the first business day before the holiday.

Preceding
Choose the first business day before the given holiday.

ModifiedPreceding
Choose the first business day before the given holiday unless it belongs to a different month, in which case choose the first business day after the holiday.

Unadjusted
Do not adjust.

data Calendar

Calendar definitions

Constructors

NullCalendar
No holidays

WeekendsOnly
Only weekends are holidays

TARGET
TARGET calendar, see http://www.ecb.int

JointCalendar Calendar Calendar
Calendar composed by two calendars

adjust :: Calendar -> Date -> BusinessDayConvention -> Date

Adjust date w.r.t. given calendar and business day convention

advanceByPeriod :: Calendar -> Date -> Period -> BusinessDayConvention -> Bool -> Date

Same as advanceByUnits, only the period is given directly as a period, not in terms of units.

advanceByUnits :: Calendar -> Date -> Int -> Unit -> BusinessDayConvention -> Bool -> Date

Advances a date w.r.t. a given calendar and business day convention by a given period If end of month (eom) is set to True an end of month date will be also an end of month date after the shift.

endOfMonth :: Calendar -> Date -> Date

Last business day in a month for a given date w.r.t. a given calendar

isBusinessDay :: Calendar -> Date -> Bool

True if date is a business day w.r.t. a given calendar

isEndOfMonth :: Calendar -> Date -> Bool

True if date is the last business day in month w.r.t. a given calendar

isHoliday :: Calendar -> Date -> Bool

True if date is a holiday w.r.t. a given calendar