Time.Date

Description

This module contains the definition of the date type and associated functions.

Documentation

newtype Date

Type representing dates in terms of a serial number.

Constructors

Date Int

date :: Int -> Int -> Int -> Date

Construct a date from given day, month, year

dayOfMonth :: Date -> Int

Day index within a month

dayOfYear :: Date -> Int

Day index within a year

isLeapYear :: Int -> Bool

True if given year is a leap year

minus :: Date -> Int -> Date

Subtract an integer from a date's serial number

minusPeriod :: Date -> Period -> Date

Subtract a period from a date

month :: Date -> Int

Month of a date

monthLength :: Int -> Bool -> Int

Length of month. If leap is True February is considered to have 29 days.

plus :: Date -> Int -> Date

Add an integer to a date's serial number

plusPeriod :: Date -> Period -> Date

Add a period to a date

serialNumber :: Date -> Int

The serial number of a date

weekday :: Date -> Int

Weekday of a date, where 1 represents Sunday, 2 Monday, ... , 7 Saturday

year :: Date -> Int

Year of a date