Time.Schedule

Description

This module contains the schedule type definition and helper functions to construct schedules. It also contains standard conventions for schedule constructions.

Documentation

type Schedule = [Date]

Schedule definition, it is nothing more than a list of dates

data ScheduleConvention

Schedule convention data type. Summarizes conventions needed for schedule construction.

eurSwapFixLegConv :: ScheduleConvention

Eur swap fixed leg standard conventions

eurSwapFloatLegConv :: ScheduleConvention

Eur swap float leg standard conventions

makeSchedule :: ScheduleConvention -> Date -> Date -> Schedule

Simplified schedule generation by convention, effective and termination date

schedule

Arguments

:: Date
Effective date

-> Date
Termination date

-> Period
Period between dates

-> Calendar
Calendar for date adjustment

-> BusinessDayConvention
Business day convention for date adjustment

-> BusinessDayConvention
Business day convention for termination date adjustment

-> DateGenerationRule
Rule for schedule generation

-> Bool
End of month flag

-> Maybe Date
First date (if not equal to effective date)

-> Maybe Date
Next to last date (if not equal to termination date)

-> Schedule
Resulting schedule

Schedule construction by effective and termination date, period, calendar, business day convention, business day convention for termination date, date generation rule, end of month flag, first date and next to last date